+++ ProcessType.cs
@@ -1,6 +1,6 @@
 ﻿namespace ArcelorMittal.UnifiedWeightSystem.Common.Sites
 {
-    public enum ProcessType : byte
+    public enum ProcessType : int
     {
         ManualStatic = 0, // Статика
         ManualDynamic = 1, // Динамика (без распознавания)
